+++ src/TodoErrors/TodoErrors.tsx
@@ -1,0 +1,26 @@
+import React from 'react';
+
+interface TodoErrorsProps {
+  errorMessage: string;
+  setErrorMessage: (value: string) => void;
+}
+
+export const TodoErrors: React.FC<TodoErrorsProps> = ({
+  errorMessage,
+  setErrorMessage,
+}) => {
+  return (
+    <div
+      data-cy="ErrorNotification"
+      className={`notification is-danger is-light has-text-weight-normal ${errorMessage ? '' : 'hidden'}`}
+    >
+      <button
+        data-cy="HideErrorButton"
+        type="button"
+        className="delete"
+        onClick={() => setErrorMessage('')}
+      />
+      {errorMessage}
+    </div>
+  );
+};
